+++ ionoscloud/resource_server.go
@@ -216,21 +216,11 @@ func resourceServer() *schema.Resource {
 							Type:     schema.TypeBool,
 							Optional: true,
 						},
-
-						"ip": {
-							Type:     schema.TypeString,
-							Optional: true,
-							DiffSuppressFunc: func(k, old, new string, d *schema.ResourceData) bool {
-								if new == "" {
-									return true
-								}
-								return false
-							},
-						},
 						"ips": {
 							Type:     schema.TypeList,
 							Elem:     &schema.Schema{Type: schema.TypeString},
 							Computed: true,
+							Optional: true,
 						},
 						"nat": {
 							Type:     schema.TypeBool,
@@ -272,16 +262,6 @@ func resourceServer() *schema.Resource {
 									},
 									"target_ip": {
 										Type:     schema.TypeString,
-										Optional: true,
-									},
-									"ip": {
-										Type:     schema.TypeString,
-										Optional: true,
-										Computed: true,
-									},
-									"ips": {
-										Type:     schema.TypeList,
-										Elem:     &schema.Schema{Type: schema.TypeString},
 										Optional: true,
 									},
 									"port_range_start": {
@@ -586,10 +566,17 @@ func resourceServerCreate(ctx context.Context, d *schema.ResourceData, meta inte
 		nic.Properties.FirewallActive = boolAddr(d.Get("nic.0.firewall_active").(bool))
 		nic.Properties.Nat = boolAddr(d.Get("nic.0.nat").(bool))
 
-		if v, ok := d.GetOk("nic.0.ip"); ok {
-			ips := strings.Split(v.(string), ",")
-			if len(ips) > 0 {
-				nic.Properties.Ips = &ips
+		if v, ok := d.GetOk("nic.0.ips"); ok {
+			raw := v.([]interface{})
+			if raw != nil && len(raw) > 0 {
+				ips := make([]string, 0)
+				for _, rawIp := range raw {
+					ip := rawIp.(string)
+					ips = append(ips, ip)
+				}
+				if ips != nil && len(ips) > 0 {
+					nic.Properties.Ips = &ips
+				}
 			}
 		}
 
@@ -900,7 +887,7 @@ func resourceServerRead(ctx context.Context, d *schema.ResourceData, meta interf
 		setPropWithNilCheck(network, "mac", nic.Properties.Mac)
 
 		if nic.Properties.Ips != nil && len(*nic.Properties.Ips) > 0 {
-			network["ip"] = (*nic.Properties.Ips)[0]
+			network["ips"] = *nic.Properties.Ips
 		}
 
 		if firewallId, ok := d.GetOk("firewallrule_id"); ok {
@@ -1121,10 +1108,17 @@ func resourceServerUpdate(ctx context.Context, d *schema.ResourceData, meta inte
 			properties.Name = &vStr
 		}
 
-		if v, ok := d.GetOk("nic.0.ip"); ok {
-			ips := strings.Split(v.(string), ",")
-			if len(ips) > 0 {
-				properties.Ips = &ips
+		if v, ok := d.GetOk("nic.0.ips"); ok {
+			raw := v.([]interface{})
+			if raw != nil && len(raw) > 0 {
+				ips := make([]string, 0)
+				for _, rawIp := range raw {
+					ip := rawIp.(string)
+					ips = append(ips, ip)
+				}
+				if ips != nil && len(ips) > 0 {
+					properties.Ips = &ips
+				}
 			}
 		}