+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 5.2.3
 
 - issue #43 - new imports for volume, user, group, share, IPfailover and loadbalancer
+- issue #47 - corrected nic resource to accept a list of strings for ips parameter 
 
 ## 5.2.2